% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coexpnet.R
\name{Build_Coexpnet}
\alias{Build_Coexpnet}
\title{A parallel co-expression network builder}
\usage{
Build_Coexpnet(
  parent_expr,
  child_expr,
  c_names,
  p_names,
  method = c("univariate", "joint"),
  k = NULL,
  nthreads = NULL
)
}
\arguments{
\item{parent_expr}{a sample (rows) x gene (columns) continuous expression profile matrix containing 
the the parent genes. This matrix also represents one half of all desirable co-expression
gene pairs.}

\item{child_expr}{a sample (rows) x gene (columns) continuous expression profile matrix containing 
the the child genes. This matrix also represents the other half of all desirable co-expression
gene pairs. The parent_expr and child_expr can be the same expression matrix if
a complete graph between all genes in a dataset is required.}

\item{c_names}{a vector containing names of the child genes.}

\item{p_names}{a vector containing names of the parent genes. The p_names and c_names
vector can be the same if a parent_expr and child_expr are equivalent.}

\item{method}{a string specifying which discretization method to use. The options are:
'univariate' that will choose \code{Ckmeans.1d.dp} or 'joint' that will choose
\code{GridOnCluster}.}

\item{k}{a range of clusters to be found on \code{parent_expr} and \code{child_expr}. 
The default minimum is 2 and the maximum is automatically computed to require atleast 
five points per cluster. If the \code{method} picked is \code{'joint'}, then a \code{Silhouette} 
score determines the number of \code{k}; if the \code{method} picked is \code{'univariate'}, 
then a \code{BIC} score determines the number of \code{k}.}

\item{nthreads}{an integer specifying the number of parallel threads to be used. If not specified,
a non-parallel version of the code would be executed.}
}
\value{
A data frame containing information on all possible parent x child co-expression pairs. 
The following items are reported about each gene pair:
\item{PID}{the one(1)-based column index in \code{parent_expr} matrix of the parent gene.}

\item{CID}{the one(1)-based column index in \code{child_expr} matrix of the child gene.}

\item{PNAMES}{the name of the parent gene.}

\item{CNAMES}{the name of the child gene.}

\item{PVALUE}{the adjusted chi-squared p-value, depicting the statistical dependency between the gene pair.}

\item{ESTIMATE}{the cramer's v effect size, depicting the strength of the relationship between the 
gene pair.}
}
\description{
Builds a co-expression network based on the Pearson's chi-square test. 
The builder computes a chi-squared p-value and a cramer's V effect size for each pair of gene.
The builder provides options to discretize the expression profiles by either Ckmeans.1d.dp,
that is an optimal univariate clustering algorithm, or GridOnCluster, that discretizes genes
involved in a co-expression pattern based on their joint distribution.
}
\seealso{
See \pkg{Ckmeans.1d.dp} for discretizing univariate continuous data.
See \pkg{GridOnClusters} for joint discretization of continuous data.
}
\author{
Ruby Sharma, Sajal Kumar and Mingzhou Song
}
