% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffCoexpnet.R
\name{Build_DiffCoexpnet}
\alias{Build_DiffCoexpnet}
\title{A parallel differential co-expression network builder}
\usage{
Build_DiffCoexpnet(
  exp_matr,
  n_conditions,
  conditions,
  indices,
  g_names,
  method = c("univariate", "joint"),
  k = NULL,
  nthreads = NULL
)
}
\arguments{
\item{exp_matr}{a sample (rows) x gene (columns) continuous expression profile matrix representing 
gene expression at given samples. This is a combined expression profile across multiple experimental
conditions.}

\item{n_conditions}{an integer specifying the number of unique experimental conditions.}

\item{conditions}{a vector of the same size as the number of samples in \code{expr_matr}, containing (1 based)
experimental condition ids for each sample.}

\item{indices}{a matrix of all gene interactions to study. The matrix containins the indices (1 based) 
of genes in \code{expr_matr}. The first column represents the parent genes and the second column represents 
child genes.}

\item{g_names}{a vector containing names of all genes in \code{expr_matr}.}

\item{method}{a string specifying which discretization method to use. The options are:
'univariate' that will choose \code{Ckmeans.1d.dp} or 'joint' that will choose
\code{GridOnCluster}.}

\item{k}{a range of clusters to be found on \code{exp_matr}. The default minimum is 2 and the 
maximum is automatically computed to require atleast five points per cluster per experimental condition. 
If the \code{method} picked is \code{'joint'}, then a \code{Silhouette} score determines the number of 
\code{k}; if the \code{method} picked  is \code{'univariate'}, then a \code{BIC} score determines the 
number of \code{k}.}

\item{nthreads}{an integer specifying the number of parallel threads to be used. If not specified,
a non-parallel version of the code would be executed.}
}
\value{
A data frame containing information on all possible parent x child differential co-expression pairs. 
The following items are reported about each gene pair:
\item{PID}{the one(1)-based column index in \code{expr_matr} matrix of the parent gene.}

\item{CID}{the one(1)-based column index in \code{expr_matr} matrix of the child gene.}

\item{PNAMES}{the name of the parent gene.}

\item{CNAMES}{the name of the child gene.}

\item{PVALUE}{the adjusted sharma-song chi-squared p-value, depicting the amount of interaction 
heterogeneity in the gene pair across the experimental conditions.}

\item{ESTIMATE}{the sharma-song effect size, depicting the amount of interaction heterogeneity in
the gene pair across the experimental conditions.}
}
\description{
Builds a differential co-expression network based on the Sharma-Song test. 
The builder computes a Sharma-Song chi-squared p-value and an effect size for each pair of gene,
under different experimental conditions.The builder provides options to discretize the expression 
profiles by either Ckmeans.1d.dp, that is an optimal univariate clustering algorithm, or 
GridOnCluster, that discretizes genes involved in co-expression patterns across the conditions, 
based on their joint distribution.
}
\seealso{
See \pkg{Ckmeans.1d.dp} for discretizing univariate continuous data.
See \pkg{GridOnClusters} for joint discretization of continuous data.
}
\author{
Ruby Sharma, Sajal Kumar and Mingzhou Song
}
