% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{DiscretizeData}
\alias{DiscretizeData}
\title{Parallel discretization of each column in 'data'}
\usage{
DiscretizeData(data, ncores, levls = -1, n_cond = 1)
}
\arguments{
\item{data}{A samples x genes normalized matrix.}

\item{ncores}{Number of parallel threads to use.}

\item{levls}{Number of levels for each column in 'data' (optional)}

\item{n_cond}{Number of conditions (groups) samples of 'data' represent (optional)}
}
\value{
A matrix containing the discretized \code{data}.
}
\description{
Parallel discretization of each column in 'data'
}
\seealso{
See \pkg{Ckmeans.1d.dp} for discretizing univariate continuous data.
}
\author{
Ruby Sharma, Sajal Kumar and Mingzhou Song
}
