% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{FilterGenesByMAD}
\alias{FilterGenesByMAD}
\title{Returns indices of low variance genes using MAD}
\usage{
FilterGenesByMAD(data, nthreads = -1)
}
\arguments{
\item{data}{a sample (rows) x gene (columns) continuous expression profile matrix.}

\item{nthreads}{an integer specifying the number of parallel threads to be used. If not specified,
a non-parallel version of the code would be executed.}
}
\value{
A logical vector of size ncol(data) containing whether a particular gene show be kept or removed.
}
\description{
Employs Median Absolute Deviation (MAD) to return genes expression profiles with zero MAD or their
MAD among the bottom 5%.
}
\author{
Ruby Sharma, Sajal Kumar and Mingzhou Song
}
