% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{GetInteractionIndices}
\alias{GetInteractionIndices}
\title{Returns common interaction to be considered for differential co-expression network.}
\usage{
GetInteractionIndices(coexpnet_list)
}
\arguments{
\item{coexpnet_list}{a list of 2 or more coexpression networks, as returned by \code{Build_Coexpnet()}.}
}
\value{
A matrix of all gene interactions to study for \code{Build_DiffCoexpnet()}. It containins the indices 
(1 based) of parent and child genes.
}
\description{
Given a list of coexpression networks, this function returns the union of common significant
interactions across all co-expression networks, to be considered for differential co-expression
network
}
\author{
Ruby Sharma, Sajal Kumar and Mingzhou Song
}
