% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{PlotDiffCoExpNet}
\alias{PlotDiffCoExpNet}
\title{Plot differential expression across multiple conditions.}
\usage{
PlotDiffCoExpNet(
  diffcoexp_res,
  exp_matr,
  n_conditions,
  conditions,
  g_names,
  clrs,
  cond_cat,
  exp_title
)
}
\arguments{
\item{diffcoexp_res}{the differential coexpression network, as returned by \code{Build_DiffCoexpnet()}.}

\item{exp_matr}{a sample (rows) x gene (columns) continuous expression profile matrix representing 
gene expression at given samples. This is a combined expression profile across multiple experimental
conditions.}

\item{n_conditions}{an integer specifying the number of unique experimental conditions.}

\item{conditions}{a vector of the same size as the number of samples in \code{expr_matr}, containing (1 based)
experimental condition ids for each sample.}

\item{g_names}{a vector containing names of all genes in \code{expr_matr}.}

\item{clrs}{a vector of size \code{n_conditions} specifying what colors to use for each experimental 
condition}

\item{cond_cat}{a vector of size \code{n_conditions} specifying the string category for each experimental
condition}

\item{exp_title}{the title for the differential experiment}
}
\description{
Given results from differential coexpression, expression profiles and a vector containing conditions id,
this function plots the differential expression across multiple conditions in the same plot.
}
\author{
Ruby Sharma, Sajal Kumar and Mingzhou Song
}
