% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharmaSongEsDist.R
\name{SharmaSongEsDist}
\alias{SharmaSongEsDist}
\title{Sharma-Song effect size distribution generator}
\usage{
SharmaSongEsDist(N, n_conditions, nthreads, d, num = 1e+05)
}
\arguments{
\item{N}{a vector of sample size for each experimental conditions.}

\item{n_conditions}{number of experimental conditions.}

\item{nthreads}{number of parallel threads to use.}

\item{d}{the maximum dimension (row and column) for simulating tables.}

\item{num}{number of simulations to consider (default:100000)}
}
\value{
A data frame containing the SharmaSong test statistic, p-value and effect size for all \code{num} 
simulations.
}
\description{
This generator simulates differential interactions (alternate population) and computes their Sharma-Song effect
size. The resultant distribution can be used to threshold effect size as well as calculate the statistical
power.
}
\seealso{
See \pkg{DiffXTables} for computing the SharmaSong test statistic for n tables.
}
\author{
Ruby Sharma, Sajal Kumar and Mingzhou Song
}
